Use a CDN for dependencies
Instead of hosting dependencies such as jQuery ourselves, we can use a
CDN. This reduces the amount of stuff we have to serve. If we use a
popular CDN, it also increases the likelihood that our dependency is
already in the users cache and they can avoid downloading it entirely.

We use a CDN for the following dependencies:
- jQuery: Google CDN (https://developers.google.com/speed/libraries)
- jQuery UI: Google CDN (https://developers.google.com/speed/libraries)
- jQuery UI Touch Punch: CDNJS (https://cdnjs.com/libraries/jqueryui-touch-punch)
- Bootstrap: Bootstrap CDN (https://www.bootstrapcdn.com/legacy/bootstrap)

This required us to bump the version slightly for jQuery, but we stayed
within the same major verison (1.x).