fix camper-not-assigned report bug
previously, the report of campers not assigned to a chug perek had a few bugs where with multiple edot and blocks/chug groups which differed by edah, it would show an individual camper as being listed in every edah, and show every chug group/block as missing an assignment (even if the camper shouldn't be assigned to that - e.g. a-side campers should not have Roo assignments, but it would show them as missing an assignment for that perek).

updated the SQL query to rectify this issue so the blocks/groups are specific to the edah and session the camper is registered for