Take attendance
Creates page to allow a chug leader to take attendance, saving the campers who are present to the database (note this necessitates the creation of a new table, "attendance_present" which keeps track of campers who show up. It will likely get quite large, so I'll probably add a "purge" button to the to-be-created admin menu. It also requires a new column in the chug_groups table tracking which block is active/current for each group)