Script to easily pull up testing database
Creates a script which can be called to easily pull up the active mysql database (without searching for the container id or selecting "camprama_chugbot_db" as the database to work in). When the Docker container isn't running, it automatically sends an error and quits.

Also fixes a minor bug in the docker https config