Leveling Autosave, Right-Click Assignments, Updates, + Fixes
When an admin is making chug assignments, they can now move campers between chugim by right-clicking the camper tile and selecting a new chug. New chugim autosave (after 3 seconds of no changes), and Bootstrap toasts are used to alert the user a save is in progress and when it completes.

Additional minor fixes:
 - Issue with all .html pages where dropdowns are nonresponsive
 - Chug counts on leveling page randomly showing the number of assigned campers as double
 - Duplicate warnings not showing
 - Improvements with handling campers not assigned a chug